LPS-75777 Move ConfigurationClusterDestinationNames into internal package
