+++ modules/apps/foundation/portal-configuration/portal-configuration-cluster/src/main/java/com/liferay/portal/configuration/cluster/internal/ConfigurationSynchronousConfigurationListener.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.configuration.cluster.internal;
 
-import com.liferay.portal.configuration.cluster.constants.ConfigurationClusterDestinationNames;
+import com.liferay.portal.configuration.cluster.internal.constants.ConfigurationClusterDestinationNames;
 import com.liferay.portal.kernel.cluster.ClusterLink;
 import com.liferay.portal.kernel.cluster.Priority;
 import com.liferay.portal.kernel.messaging.Destination;
