+++ modules/apps/foundation/portal-configuration/portal-configuration-cluster/src/main/java/com/liferay/portal/configuration/cluster/internal/constants/ConfigurationClusterDestinationNames.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.configuration.cluster.constants;
+package com.liferay.portal.configuration.cluster.internal.constants;
 
 /**
  * @author Michael C. Han
