+++ modules/apps/foundation/portal-configuration/portal-configuration-cluster/src/main/java/com/liferay/portal/configuration/cluster/internal/messaging/ConfigurationClusterConfigurator.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.configuration.cluster.internal.messaging;
 
-import com.liferay.portal.configuration.cluster.constants.ConfigurationClusterDestinationNames;
+import com.liferay.portal.configuration.cluster.internal.constants.ConfigurationClusterDestinationNames;
 import com.liferay.portal.kernel.messaging.Destination;
 import com.liferay.portal.kernel.messaging.DestinationConfiguration;
 import com.liferay.portal.kernel.messaging.DestinationFactory;
