+++ modules/apps/foundation/portal-configuration/portal-configuration-cluster/src/main/java/com/liferay/portal/configuration/cluster/internal/messaging/ConfigurationMessageListener.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.configuration.cluster.internal.messaging;
 
-import com.liferay.portal.configuration.cluster.constants.ConfigurationClusterDestinationNames;
 import com.liferay.portal.configuration.cluster.internal.ConfigurationThreadLocal;
+import com.liferay.portal.configuration.cluster.internal.constants.ConfigurationClusterDestinationNames;
 import com.liferay.portal.configuration.persistence.ReloadablePersistenceManager;
 import com.liferay.portal.kernel.messaging.BaseMessageListener;
 import com.liferay.portal.kernel.messaging.Destination;
